backend/requirements.txt
fastapi
uvicorn
pydantic
sympy
numpy
groq
python-multipart
google-generativeai
pillow
httpx
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
supabase==2.0.0
langgraph
typing_extensions
dspy-ai
huggingface-hub
